# swagger.yaml
openapi: 3.0.0
info:
  description: "This is a professional api project I am creating for our first assignment in full-stack army. This is a simple code-snippet sharing api application where anyone can create a new account, login to their system and create code snippets that will be visible publicly."
  version: "1.0.0"
  title: "SnipNest - Where Code Find It's Nest"
  contact:
    name: Safin Ahmed
    url: https://safinahmed.netlify.app/
    email: safin.ahmed2000@gmail.com
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SAFINAHMED2000/SnipNest/1.0.0
  - description: Dev Server
    url: http://localhost:4000/api/v1
...
paths:
...
  "/admin/comments":
    get:
      operationId: findAllComments
      description: "Retrieve all comments"
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/sortParam"
        - $ref: "#/components/parameters/searchParam"
      responses:
        200:
          description: "Return a list of comments"
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/CommentDTO"
                        - $ref: "#/components/schemas/Timestamp"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/admin/comments?page=2&limit=10"
                      next:
                        type: string
                        format: url
                        example: "/admin/comments?page=3&limit=10"
                      prev:
                        type: string
                        format: url
                        example: "/admin/comments?page=1&limit=10"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
      tags:
        - Comment
    post:
      operationId: createComment
      security:
        - bearerAuth: []
      tags: [Comment]
      description: Create a new comment
      requestBody:
        description: To create a new comment, the admin must need to provide valid data
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/CommentRequestBody"
      responses:
        201:
          description: Successfully created a new comment
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/CommentCreateDTO"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
...
  "/admin/stars":
    get:
      operationId: findAllStars
      description: "Retrieve all stars"
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/sortParam"
        - $ref: "#/components/parameters/searchParam"
      responses:
        200:
          description: "Return a list of stars"
          content:
            "application/json":
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/StarReactDTO"
                        - $ref: "#/components/schemas/Timestamp"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: "/admin/stars?page=2&limit=10"
                      next:
                        type: string
                        format: url
                        example: "/admin/stars?page=3&limit=10"
                      prev:
                        type: string
                        format: url
                        example: "/admin/stars?page=1&limit=10"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
      tags:
        - Star
    post:
      operationId: createStar
      security:
        - bearerAuth: []
      tags: [Star]
      description: Create a new Star
      requestBody:
        description: To create a new Star, the admin must need to provide valid data
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/StarReactRequestBody"
      responses:
        201:
          description: Successfully created a new Star React
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/StarReactCreateDTO"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: datetime
        updatedAt:
          type: string
          format: datetime
      required: [createdAt, updatedAt]
    ProfileDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        display_name:
          type: string
          example: safinahmed
        job_title:
          type: string
          example: Backend Developer
        bio:
          type: string
        follower_count:
          type: integer
          example: 100
...
    CommentDTO:
      type: object
      properties:
        id:
          type: string
          example: "asdasdad212easd"
        snippet_id:
          type: string
          example: "asdasdads21asd"
        profile:
          $ref: "#/components/schemas/ProfileDTO"
        content:
          type: string
    CommentCreateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: "Comment created successfully"
        data:
          $ref: "#/components/schemas/CommentDTO"
        links:
          type: object
          properties:
            self:
              type: string
              example: "/comments/asdasdsadasd"
            snippet:
              type: string
              example: "/snippets/asdasdasdasda"
...
    CommentRequestBody:
      type: object
      properties:
        content:
          type: string
          example: "This is a comment"
        snippet_id:
          type: string
          example: "asdasd123123asd"
        profile_id:
          type: string
          example: "sasdasdasd213123asdas"
      required: [content, snippet_id, profile_id]
    StarReactDTO:
      type: object
      properties:
        snippet:
          type: object
          properties:
            id:
              type: string
              example: "asdasdasd"
            title:
              type: string
              example: "Test Snippet"
        profile:
          type: object
          properties:
            displayName:
              type: string
              example: sniper123
            avatar:
              type: string
    StarReactCreateDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 201
        message:
          type: string
          example: "Star react created successfully"
        data:
          $ref: "#/components/schemas/StarReactDTO"
        links:
          type: object
          properties:
            self:
              type: string
              example: "/stars/asdasdsadasd"
            snippet:
              type: string
              example: "/snippets/asdasdasdasda"
    StarReactRequestBody:
      type: object
      properties:
        snippet_id:
          type: string
          example: "asdasd123123asd"
        profile_id:
          type: string
          example: "sasdasdasd213123asdas"
      required: [snippet_id, profile_id]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalPage:
          type: integer
          format: int32
          example: 5
        totalItems:
          type: integer
          format: int32
          example: 50
      required: [page, limit, totalPage, totalItems]
  responses:
    "500":
      description: "Internal Server Error"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later
    "400":
      description: "Bad Request"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad Request
              data:
                type: object
                properties: { field: { type: "string" } }
    "401":
      description: "Unauthorized"
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
...
  parameters:
    pageParam:
      in: query
      name: page
      description: current page number
      schema:
        type: integer
        format: int32
        example: 2
    limitParam:
      in: query
      name: limit
      description: maximum items to be returned
      schema:
        type: integer
        format: int32
        example: 10
    sortParam:
      in: query
      name: sort
      description: sort the result by a property in a defined order [sortProperty:sortOrder]
      schema:
        type: string
        example: "title:asc"
    searchParam:
      in: query
      name: search
      description: Search term
      schema:
        type: string
        example: title
...